Ellipsise long page titles (Fixes #255)
In page list, use full width if no action buttons are shown and limit
width when action buttons are displayed to prevent collisions between
title and action buttons.

In page title, the title is simply ellipsised for now. We don't want
to encourage people to use long page titles anyway.