Correctly calculate directory sizes when using an object store primary storage
A object store storage should use an NoopScanner since it is not
possible to scan the directory structure of an object store. This makes
sure that the collective folder mount point doesn't override the NoopScanner
that the object store storage use.

Taken over from https://github.com/nextcloud/groupfolders/pull/1630